inherit item component from show component allowing to render comment
