+++ modules/overviews/app/components/_index.sass
@@ -1,4 +1,4 @@
-@import "./overviews/project_custom_fields/item_component.sass"
+@import "./overviews/project_custom_fields/show_component.sass"
 @import "./overviews/project_custom_fields/items_component.sass"
 @import "./overviews/project_phases/item_component.sass"
 @import "./overviews/overview_grid_component.sass"
