+++ modules/overviews/app/components/overviews/project_custom_fields/show_component.html.erb
@@ -43,6 +43,18 @@
         render_calculation_error
       end
     end
+
+    if show_comment?
+      custom_field_value_container.with_row(w: :full, mt: 1) do
+        render Primer::Alpha::TextArea.new(
+          name: :custom_comment,
+          label: I18n.t("attributes.comment"),
+          value: @project_custom_field.comment_for(@project)&.text,
+          rows: 5,
+          readonly: true
+        )
+      end
+    end
   end
 %>
